fix(TaskEditorBottomSheet): resolve the conflict with develop branch.
