+++ app/src/main/java/com/giraffe/tudeeapp/presentation/screen/taskeditor/TaskEditorBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.giraffe.tudeeapp.presentation.taskeditor
+package com.giraffe.tudeeapp.presentation.screen.taskeditor
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -10,6 +10,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
@@ -22,11 +23,17 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
+import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -46,16 +53,77 @@ import com.giraffe.tudeeapp.design_system.component.button_type.PrimaryButton
 import com.giraffe.tudeeapp.design_system.component.button_type.SecondaryButton
 import com.giraffe.tudeeapp.design_system.theme.Theme
 import com.giraffe.tudeeapp.domain.entity.task.TaskPriority
+import com.giraffe.tudeeapp.presentation.utils.EventListener
+import com.giraffe.tudeeapp.presentation.utils.errorToMessage
 import com.giraffe.tudeeapp.presentation.utils.formatAsLocalizedDate
 import com.giraffe.tudeeapp.presentation.utils.toStringResource
+import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
+import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
-import kotlinx.datetime.LocalDate
+import org.koin.androidx.compose.koinViewModel
+
+@RequiresApi(Build.VERSION_CODES.O)
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun TaskEditorBottomSheet(
+    taskId: Long?,
+    onDismissRequest: () -> Unit,
+    modifier: Modifier = Modifier,
+    selectedDate: LocalDate,
+    onSuccess: (String) -> Unit = {},
+    onError: (String) -> Unit = {},
+    viewModel: TaskEditorViewModel = koinViewModel()
+) {
+    val context = LocalContext.current
+    val scope = rememberCoroutineScope()
+    val bottomSheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
+    val taskEditorUiState by viewModel.state.collectAsState()
+    LaunchedEffect(key1 = taskId, selectedDate) {
+        if (taskId == null) {
+            viewModel.setDueDate(selectedDate)
+        } else {
+            viewModel.loadTask(taskId)
+        }
+    }
+    EventListener(
+        events = viewModel.effect,
+    ) { event ->
+        when (event) {
+            TaskEditorEffect.TaskAddedSuccess -> onSuccess(context.getString(R.string.task_added_successfully))
+            is TaskEditorEffect.Error -> onError(context.errorToMessage(event.error))
+            TaskEditorEffect.TaskEditedSuccess -> onSuccess(context.getString(R.string.task_edited_successfully))
+            TaskEditorEffect.DismissTaskEditor -> onDismissRequest()
+        }
+
+    }
+
+    ModalBottomSheet(
+        onDismissRequest = {
+            viewModel.cancel()
+            onDismissRequest()
+            scope.launch { bottomSheetState.hide() }
+        },
+        sheetState = bottomSheetState,
+        modifier = modifier.fillMaxHeight(0.95f),
+        containerColor = Theme.color.surface
+    ) {
+
+        TaskEditorBottomSheetContent(
+            taskEditorState = taskEditorUiState,
+            actions = viewModel,
+            selectedDate = selectedDate,
+            isNewTask = taskId == null
+        )
+    }
+
+
+}
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
-fun TaskEditorBottomSheetContent(
+private fun TaskEditorBottomSheetContent(
     taskEditorState: TaskEditorState,
     selectedDate: LocalDate = Clock.System.now().toLocalDateTime(TimeZone.UTC).date,
     actions: TaskEditorInteractionListener,
